Refactoring model. step 2 (Tests, Remame)
