+++ Controllers/solar_viewer_controller.py
