--- Models/timeLineModel.py
@@ -1,13 +1,0 @@
-class TimeLineModel:
-    def __init__(self):
-        self.__observers = []
-
-    def addObserver(self, inObserver):
-        self.__observers.append(inObserver)
-
-    def removeObserver(self, inObserver):
-        self.__observers.remove(inObserver)
-
-    def notifyObserver(self):
-        for x in self.__observers:
-            x.modelIsChanged()
